# src/main/resources/api/dev-practice.openapi.yaml
openapi: 3.0.3
info:
  title: La Forge API
  version: 0.1.0
  description: |
    Contrat du prototype de laboratoire d’entraînement technique. Texte métier en français, identifiants API en anglais.

    Périmètre : profil, catalogue, exercices versionnés, tentatives, indices, corrigés, révisions, veille, défis collectifs, édition et tuteur IA optionnel. Pas d’exécution distante de code, génération automatique de contenus, classement RH ou gestion multi-entreprises. Une installation correspond à une entreprise ou un groupe pilote. Les rôles sont gérés hors API.

    Authentification : compte créé localement via /auth/registrations (email + mot de passe), activé par confirmation
    du jeton de vérification reçu par courriel (/auth/verifications). Connexion via /auth/sessions : le serveur émet
    lui-même un bearer JWT de courte durée (signature RS256) accompagné d’un jeton de renouvellement opaque
    (/auth/sessions/refresh). Le serveur valide signature, issuer, audience et expiration ; utilisateur courant déduit
    de sub, jamais du corps de requête. Rôles LEARNER, AUTHOR, REVIEWER, ADMIN cumulatifs ; tous les comptes peuvent
    pratiquer dès activation. x-required-roles signifie au moins un rôle de la liste, plus les contraintes de propriété
    décrites. ADMIN n’accorde pas l’accès aux tentatives privées des autres.

...

...

...

...

...

...

...
servers:
- url: /api/v1
  description: Même origine que le frontend ; préfixe de la v1.
security:
- bearerAuth: []
...
paths:
  /auth/registrations:
    post:
      tags:
      - Auth
      operationId: registerAccount
      summary: Créer un compte
      description: Crée un compte en attente d’activation et envoie un courriel de confirmation contenant un jeton
        à usage unique valable 24 h. Réponse 202 constante que l’adresse existe déjà ou non, afin de ne jamais
        révéler l’existence d’un compte. Aucune authentification requise.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationInput'
      responses:
        '202':
          description: Demande d’inscription acceptée ; courriel envoyé si l’adresse est éligible.
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/verifications:
    post:
      tags:
      - Auth
      operationId: verifyAccount
      summary: Confirmer le compte
      description: Valide le jeton reçu par courriel, active le compte et attribue le rôle LEARNER. Jeton à usage
        unique, expiré après 24 h ou après utilisation.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationInput'
      responses:
        '200':
          description: Compte activé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/verifications/resend:
    post:
      tags:
      - Auth
      operationId: resendVerification
      summary: Renvoyer le courriel de confirmation
      description: Renvoie un nouveau jeton de vérification si le compte existe et n’est pas déjà actif. Réponse
        202 constante, limitée en débit par adresse.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationInput'
      responses:
        '202':
          description: Demande acceptée ; courriel renvoyé si éligible.
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/sessions:
    post:
      tags:
      - Auth
      operationId: createSession
      summary: Se connecter
      description: Authentifie par email et mot de passe un compte activé, émet un jeton d’accès JWT de courte durée
        et un jeton de renouvellement opaque.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Session créée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
      - Auth
      operationId: revokeSession
      summary: Se déconnecter
      description: Révoque le jeton de renouvellement fourni. Idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Session révoquée.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/sessions/refresh:
    post:
      tags:
      - Auth
      operationId: refreshSession
      summary: Renouveler la session
      description: Échange un jeton de renouvellement valide contre un nouveau jeton d’accès et un nouveau jeton
        de renouvellement (rotation). La réutilisation d’un jeton déjà consommé révoque toute la famille de jetons.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Session renouvelée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /attempts/{attemptId}/tutor-exchanges:
    get:
      tags:
      - Tutor
      operationId: listTutorExchanges
      summary: Lire mes échanges avec le tuteur
      description: Propriétaire uniquement, tentative SUBMITTED requise. Historique disponible même si IA désactivée.
        Ordre createdAt croissant puis id.
      responses:
        '200':
          description: Lire mes échanges avec le tuteur.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
    post:
      tags:
      - Tutor
      operationId: askTutor
      summary: Poser une question après le débrief
      description: Optionnel. Propriétaire et SUBMITTED requis. Réponse JSON synchrone, délai serveur maximal 30
        s puis 503 ; aucun streaming dans ce contrat. Un échange complet est persisté atomiquement après succès.
        Maximum une génération simultanée par tentative (409). Contexte limité à l’exercice, au corrigé, à cette
        réponse et à cet historique ; aucun accès aux données des collègues. Ne modifie aucune évaluation.
      responses:
        '201':
          description: Poser une question après le débrief.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorExchange'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
      parameters:
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TutorQuestion'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton du fournisseur d’identité configuré. Acquisition hors périmètre.
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Nombre maximal d’éléments.
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 1000
      description: Curseur opaque retourné par la page précédente.
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 'UUID généré par le client, unique pour utilisateur + méthode + chemin. Conservation 24 h. Même
        clé et même corps : réponse originale sans répéter l’effet. Autre corps : 409. Déduplication atomique, y
        compris des appels IA. Opération encore en cours : 409 INVALID_STATE, réessayer avec la même clé.'
  responses:
    BadRequest:
      description: Requête syntaxiquement invalide.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthenticated:
      description: Jeton absent, expiré ou invalide.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        WWW-Authenticate:
          description: Challenge Bearer.
          schema:
            type: string
          example: Bearer
    Forbidden:
      description: Rôle insuffisant ou déblocage pédagogique non acquis.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Ressource inexistante ou non accessible dans le périmètre de l’appelant.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Transition interdite, tentative déjà finalisée, version de base obsolète ou conflit d’idempotence.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationFailed:
      description: Données ou contraintes métier invalides ; violations détaille les champs.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    TooManyRequests:
      description: Quota dépassé ; respecter Retry-After.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          description: Délai minimal en secondes avant nouvel essai, si disponible.
          schema:
            type: integer
            format: int32
            minimum: 0
    InternalError:
      description: Erreur serveur inattendue.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServiceUnavailable:
      description: Assistance IA désactivée ou indisponible ; aucun échange partiel enregistré.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          description: Délai minimal en secondes avant nouvel essai, si disponible.
          schema:
            type: integer
            format: int32
            minimum: 0
  schemas:
    RegistrationInput:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 12
          maxLength: 200
          description: Vérifié côté serveur ; jamais retourné ni journalisé en clair.
        displayName:
          type: string
          maxLength: 200
          minLength: 1
      required:
      - email
      - password
      - displayName
      description: Réponse toujours 202, y compris si l’adresse est déjà utilisée, pour ne pas révéler l’existence
        d’un compte.
    VerificationInput:
      type: object
      additionalProperties: false
      properties:
        token:
          type: string
          maxLength: 200
          minLength: 1
      required:
      - token
    ResendVerificationInput:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 320
      required:
      - email
    LoginInput:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 1
          maxLength: 200
      required:
      - email
      - password
    Session:
      type: object
      additionalProperties: false
      properties:
        accessToken:
          type: string
          maxLength: 8000
        refreshToken:
          type: string
          maxLength: 500
        tokenType:
          type: string
          enum:
          - Bearer
        expiresInSeconds:
          type: integer
          format: int32
          minimum: 1
        user:
          $ref: '#/components/schemas/User'
      required:
      - accessToken
      - refreshToken
      - tokenType
      - expiresInSeconds
      - user
    RefreshRequest:
      type: object
      additionalProperties: false
      properties:
        refreshToken:
          type: string
          maxLength: 500
          minLength: 1
      required:
      - refreshToken
...
    Difficulty:
      type: string
      enum:
      - BEGINNER
      - INTERMEDIATE
      - ADVANCED
    Problem:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          format: uri-reference
          example: urn:dev-practice:problem:invalid-state
        title:
          type: string
          maxLength: 200
          minLength: 1
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        detail:
          type: string
          maxLength: 4000
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          enum:
          - BAD_REQUEST
          - UNAUTHENTICATED
          - FORBIDDEN
          - NOT_FOUND
          - INVALID_STATE
          - IDEMPOTENCY_CONFLICT
          - STALE_VERSION
          - VALIDATION_FAILED
          - PRECONDITION_REQUIRED
          - RATE_LIMITED
          - AI_UNAVAILABLE
          - INTERNAL_ERROR
        traceId:
          type: string
          maxLength: 100
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
          maxItems: 100
      required:
      - type
      - title
      - status
      - code
      example:
        type: urn:dev-practice:problem:invalid-state
        title: Tentative déjà soumise
        status: 409
        code: INVALID_STATE
        detail: La réponse définitive ne peut plus être modifiée.
    Violation:
      type: object
      additionalProperties: false
      properties:
        field:
          type: string
          maxLength: 200
          minLength: 1
        message:
          type: string
          maxLength: 4000
      required:
      - field
      - message
    StackEntry:
      type: object
      additionalProperties: false
      properties:
        technology:
          type: string
          maxLength: 60
          minLength: 1
          example: Java
        currentVersion:
          type: string
          maxLength: 40
          minLength: 1
          example: '17'
        targetVersion:
          type: string
          maxLength: 40
          minLength: 1
          example: '21'
      required:
      - technology
      - currentVersion
    Preferences:
      type: object
      additionalProperties: false
      properties:
        stack:
          type: array
          items:
            $ref: '#/components/schemas/StackEntry'
          maxItems: 20
        topicIds:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 50
          uniqueItems: true
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        sessionMinutes:
          type: integer
          format: int32
          minimum: 5
          maximum: 30
        locale:
          type: string
          enum:
          - fr
          - en
        timeZone:
          type: string
          maxLength: 100
          example: Europe/Paris
          description: Identifiant IANA validé par le serveur.
      required:
      - stack
      - topicIds
      - difficulty
      - sessionMinutes
      - locale
      - timeZone
    User:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
          maxLength: 200
          minLength: 1
        roles:
          type: array
          items:
            type: string
            enum:
            - LEARNER
            - AUTHOR
            - REVIEWER
            - ADMIN
          maxItems: 4
          uniqueItems: true
        preferences:
          $ref: '#/components/schemas/Preferences'
      required:
      - id
      - displayName
      - roles
      - preferences
...
    Source:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
        url:
          type: string
          format: uri
          maxLength: 2048
        accessedAt:
          type: string
          format: date-time
      required:
      - title
      - url
      - accessedAt
...
    TutorQuestion:
      type: object
      additionalProperties: false
      properties:
        question:
          type: string
          maxLength: 4000
          minLength: 1
      required:
      - question
    TutorExchange:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        question:
          type: string
          maxLength: 4000
        answerMarkdown:
          type: string
          maxLength: 20000
          minLength: 1
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
          maxItems: 20
        createdAt:
          type: string
          format: date-time
        generatedByAi:
          type: boolean
          enum:
          - true
      required:
      - id
      - question
      - answerMarkdown
      - sources
      - createdAt
      - generatedByAi
      description: Explication indicative générée par IA, ne modifie ni le corrigé validé ni le résultat objectif.
        Sources limitées aux références de l’exercice ; tableau vide si aucune citation vérifiable.
...
    TutorPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TutorExchange'
          maxItems: 100
        nextCursor:
          type: string
          maxLength: 1000
      required:
      - items
      description: Pagination par curseur opaque ; nextCursor absent à la dernière page. Aucun total garanti. Un
        curseur doit être réutilisé avec les mêmes filtres.
...